chore: refactor code for less duplication
